requirements.txt
aniso8601==1.2.0
asn1crypto==1.3.0
boto3==1.13.16
botocore==1.16.16
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.1.4
docutils==0.15.2
Flask==0.12.1
Flask-Ask==0.9.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
pycparser==2.20
pyOpenSSL==17.0.0
python-dateutil==2.8.1
PyYAML==3.12
requests==2.23.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.25.8
Werkzeug==0.14.1
gunicorn==20.0.4
coreapi==2.3.3
beautifulsoup4==4.7.1